x Visual C++ 2003 patch
git-svn-id: svn://svn.code.sf.net/p/mediainfo/code/MediaInfoLib/trunk@1704 45c5ae6f-87cc-4fd0-92ee-e4f023fd80da

+++ Source/MediaInfo/File__Analyze_Buffer.cpp
@@ -953,9 +953,9 @@ void File__Analyze::Get_SE(int32s &Info, const char* Name)
     while(BS->Remain()>0 && BS->Get(1)==0)
         LeadingZeroBits++;
     INTEGRITY(LeadingZeroBits<=32, "(Problem)", 0)
-    double InfoD=pow(2, (float)LeadingZeroBits)-1+BS->Get(LeadingZeroBits);
+    double InfoD=pow((float)2, (float)LeadingZeroBits)-1+BS->Get(LeadingZeroBits);
     INTEGRITY(InfoD<int32u(-1), "(Problem)", 0)
-    Info=(int32s)(pow(-1, InfoD+1)*(int32u)ceil(InfoD/2));
+    Info=(int32s)(pow((double)-1, InfoD+1)*(int32u)ceil(InfoD/2));
 
     if (Config_Details>0)
         Param(Name, Info);
@@ -971,7 +971,7 @@ void File__Analyze::Skip_SE(const char* Name)
     if (Config_Details>0)
     {
         INTEGRITY(LeadingZeroBits<=32, "(Problem)", 0)
-        double InfoD=pow(2, (float)LeadingZeroBits)-1+BS->Get(LeadingZeroBits);
+        double InfoD=pow((float)2, (float)LeadingZeroBits)-1+BS->Get(LeadingZeroBits);
         INTEGRITY(InfoD<int32u(-1), "(Problem)", 0)
         Param(Name, (int32s)(pow(-1, InfoD+1)*(int32u)ceil(InfoD/2)));
     }
