+++ Source/MediaInfo/File__Analyze_Buffer_MinimizeSize.cpp
@@ -743,9 +743,9 @@ void File__Analyze::Get_SE(int32s &Info)
     while(BS->Remain()>0 && BS->Get(1)==0)
         LeadingZeroBits++;
     INTEGRITY(LeadingZeroBits<=32, "(Problem)", 0)
-    double InfoD=pow(2, (float)LeadingZeroBits)-1+BS->Get(LeadingZeroBits);
+    double InfoD=pow((float)2, (float)LeadingZeroBits)-1+BS->Get(LeadingZeroBits);
     INTEGRITY(InfoD<int32u(-1), "(Problem)", 0)
-    Info=(int32s)(pow(-1, InfoD+1)*(int32u)ceil(InfoD/2));
+    Info=(int32s)(pow((double)-1, InfoD+1)*(int32u)ceil(InfoD/2));
 }
 
 //---------------------------------------------------------------------------
@@ -766,7 +766,7 @@ void File__Analyze::Get_UE(int32u &Info)
     while(BS->Remain()>0 && BS->Get(1)==0)
         LeadingZeroBits++;
     INTEGRITY(LeadingZeroBits<=32, "(Problem)", 0)
-    double InfoD=pow(2, (float)LeadingZeroBits);
+    double InfoD=pow((float)2, (float)LeadingZeroBits);
     Info=(int32u)InfoD-1+BS->Get(LeadingZeroBits);
 }
